Refactor Gradle Build and Bump Dependencies
