+++ app/src/main/java/com/androiders/dateme/OnBoardingUiModel.kt
@@ -1,12 +1,12 @@
 package com.androiders.dateme
 
 import androidx.compose.ui.graphics.Color
-import com.androiders.dateme.ui.theme.ColorBlue
+import com.androiders.dateme.core.ui.theme.ColorBlue
 
-data class OnBoardingData(
+data class OnBoardingUiModel(
     val image: Int,
     val title: String,
     val desc: String,
     val backgroundColor: Color,
-    val mainColor: Color = ColorBlue
+    val mainColor: Color = ColorBlue,
 )
