+++ app/src/main/java/com/androiders/dateme/core/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.androiders.dateme.ui.theme
+package com.androiders.dateme.core.ui.theme
 
 import androidx.compose.ui.graphics.Color
