+++ app/src/main/java/com/androiders/dateme/core/ui/theme/Shape.kt
@@ -1,4 +1,4 @@
-package com.androiders.dateme.ui.theme
+package com.androiders.dateme.core.ui.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Shapes
