+++ app/src/main/java/com/androiders/dateme/core/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.androiders.dateme.ui.theme
+package com.androiders.dateme.core.ui.theme
 
 import androidx.compose.material.Typography
 import androidx.compose.ui.text.TextStyle
